pyproject.toml
[project]
name = "fast-ai"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.2",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.11",
]

[tool.mypy]
strict = true
ignore_missing_imports = true